requirements.txt
Flask
tweepy
python-dotenv
spoonacular
python-dotenv
requests